订单修改和优惠券修改

+++ src/main/java/org/crown/projects/classify/service/impl/OrderServiceImpl.java
@@ -115,45 +115,49 @@ public class OrderServiceImpl extends BaseServiceImpl<OrderMapper, Order> implem
         BigDecimal price = orderPARM.getPrice();
         Integer couponId = orderPARM.getCouponId();
         ProductDTO product = orderPARM.getProduct();
-        BigDecimal postFee1 = queryPostFee(product.getNum(), address.getId(), product.getId(), price);
-        if (postFee.compareTo(postFee1) != 0) {
-            log.error("邮费不正确：【传入邮费：" + postFee + ",运算出的邮费：" + postFee1 + " 】");
-            return null;
-        }
-        ProductPrice productPrice = productPriceService.query().eq(ProductPrice::getPid, product.getId()).ge(ProductPrice::getENum, product.getNum()).le(ProductPrice::getSNum, product.getNum()).entity(e -> e);
-        if (productPrice.getPrice().compareTo(product.getPrice()) != 0) {
-            log.error("商品价格不正确:【传入价格：" + product.getPrice() + ",运算出的价格：" + productPrice.getPrice() + " 】");
-            return null;
-        }
-        BigDecimal add = productPrice.getPrice().multiply(new BigDecimal(product.getNum() + "")).add(postFee1);
         //创建订单
         Order order = new Order();
-        //设置优惠券id
-        if (Objects.nonNull(couponId)) {
-            Coupon coupon = couponService.getById(couponId);
-            if (coupon.getType() != 3) {
-                add = add.subtract(coupon.getDiscount());
-            }else {
-                add = add.multiply(coupon.getDiscount());
+        if (orderPARM.getOrderType() != 1) {
+            BigDecimal postFee1 = queryPostFee(product.getNum(), address.getId(), product.getId(), price);
+            if (postFee.compareTo(postFee1) != 0) {
+                log.error("邮费不正确：【传入邮费：" + postFee + ",运算出的邮费：" + postFee1 + " 】");
+                return null;
             }
-            //设置订单优惠券
+            ProductPrice productPrice = productPriceService.query().eq(ProductPrice::getPid, product.getId()).ge(ProductPrice::getENum, product.getNum()).le(ProductPrice::getSNum, product.getNum()).entity(e -> e);
+            if (productPrice.getPrice().compareTo(product.getPrice()) != 0) {
+                log.error("商品价格不正确:【传入价格：" + product.getPrice() + ",运算出的价格：" + productPrice.getPrice() + " 】");
+                return null;
+            }
+            BigDecimal add = productPrice.getPrice().multiply(new BigDecimal(product.getNum() + "")).add(postFee1);
+
+            //设置优惠券id
+            if (Objects.nonNull(couponId)) {
+                Coupon coupon = couponService.getById(couponId);
+                if (coupon.getType() != 3) {
+                    add = add.subtract(coupon.getDiscount());
+                }else {
+                    add = add.multiply(coupon.getDiscount());
+                }
+                //设置订单优惠券
+                if (price.compareTo(add) != 0) {
+                    log.error("商品的总价格不正确：【传入价格：" + price + "，运算出的价格：" + add + " 】");
+                    return null;
+                }
+                order.setCouponId(couponId);
+                List<CouponCustomer> list = couponCustomerService.query().eq(CouponCustomer::getCouponId, couponId).eq(CouponCustomer::getOpenId, customer.getOpenId()).list();
+                for (int i = 0; i < 1; i++) {
+                    //优惠券数量减少
+                    couponService.deleteCouponCustomerById(list.get(i).getId());
+                }
+                coupon.setUsed(coupon.getUsed() + 1);
+                couponService.updateById(coupon);
+            }
             if (price.compareTo(add) != 0) {
                 log.error("商品的总价格不正确：【传入价格：" + price + "，运算出的价格：" + add + " 】");
                 return null;
             }
-            order.setCouponId(couponId);
-            List<CouponCustomer> list = couponCustomerService.query().eq(CouponCustomer::getCouponId, couponId).eq(CouponCustomer::getOpenId, customer.getOpenId()).list();
-            for (int i = 0; i < 1; i++) {
-                //优惠券数量减少
-                couponService.deleteCouponCustomerById(list.get(i).getId());
-            }
-            coupon.setUsed(coupon.getUsed() + 1);
-            couponService.updateById(coupon);
         }
-        if (price.compareTo(add) != 0) {
-            log.error("商品的总价格不正确：【传入价格：" + price + "，运算出的价格：" + add + " 】");
-            return null;
-        }
+
         //购买数量
         order.setNum(product.getNum());
         //设置订单类型
